// src/types/partials/SpecimenDefinitionTypeTested.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_classes so do not edit manually

import { TExtension } from '../partials/Extension';
import { TCodeableConcept } from '../partials/CodeableConcept';
import { TSpecimenDefinitionContainer } from '../partials/SpecimenDefinitionContainer';
import { TQuantity } from '../partials/Quantity';
import { TSpecimenDefinitionHandling } from '../partials/SpecimenDefinitionHandling';

export type TSpecimenDefinitionTypeTested = {
    id?: String;
    extension?: TExtension[];
    modifierExtension?: TExtension[];
    isDerived?: Boolean;
    type?: TCodeableConcept;
    preference: String;
    container?: TSpecimenDefinitionContainer;
    requirement?: String;
    retentionTime?: TQuantity;
    rejectionCriterion?: TCodeableConcept[];
    handling?: TSpecimenDefinitionHandling[];
};